+++ src/index.js
@@ -11,6 +11,7 @@ import "./styles/forms.css";
 import "./styles/menu.css";
 import "./styles/header.css";
 import "./icomoon/style.css";
+import "./components/CheckBoxes/checkbox.css";
 
 /**
  * layout components and navigational page elements
@@ -81,3 +82,40 @@ export { Labels } from "./components/Labels/Labels";
  */
 
 export { Breadcrumb } from "./components/Breadcrumb/Breadcrumb";
+
+/**
+ *Text Input Field
+ */
+
+export { TextField } from "./components/TextField/TextField";
+
+/*
+ * FormError
+ */
+
+export { FormError } from "./components/FormError/FormError";
+
+/**
+ * Form Label
+ */
+export { FormLabel } from "./components/FormLabel/FormLabel";
+
+/**
+ * Accordion
+ */
+export { AccordionForm } from "./components/AccordionForm/AccordionForm";
+
+/*
+ * Form RadioButton
+ */
+export { FormRadioButton } from "./components/FormRadioButton/FormRadioButton";
+
+/*
+ * Check box
+ */
+export { CheckBoxes } from "./components/CheckBoxes/CheckBoxes";
+
+/*
+ * Check box
+ */
+export { CheckBoxForm } from "./components/CheckBoxForm/CheckBoxForm";